+++ app/react/V2/Routes/Settings/Languages/LanguagesList.tsx
@@ -8,9 +8,7 @@ import { Row, createColumnHelper } from '@tanstack/react-table';
 import { Translate, I18NApi } from 'app/I18N';
 import { RequestParams } from 'app/utils/RequestParams';
 import { settingsAtom } from 'app/V2/atoms/settingsAtom';
-import { ConfirmationModal } from 'app/V2/Components/UI/ConfirmationModal';
-import { Button } from 'V2/Components/UI/Button';
-import { Table } from 'V2/Components/UI/Table';
+import { Button, Table, ConfirmationModal } from 'V2/Components/UI';
 import { useApiCaller } from 'V2/CustomHooks/useApiCaller';
 import { SettingsContent } from 'app/V2/Components/Layouts/SettingsContent';
 import { LanguageSchema } from 'shared/types/commonTypes';
